Remove some test interdependencies in unit/tools
Factor out BaseParameterTestCase into a helper module.
