+++ test/unit/tools/test_column_parameters.py
@@ -3,7 +3,7 @@ test_select_parameters.py.
 """
 from galaxy import model
 from galaxy.util import bunch
-from .test_parameter_parsing import BaseParameterTestCase
+from .util import BaseParameterTestCase
 from ..tools_support import datatypes_registry
