+++ test/unit/tools/test_select_parameters.py
@@ -1,7 +1,7 @@
 from galaxy import model
 from galaxy.tools.parameters import basic
 from galaxy.util import bunch
-from .test_parameter_parsing import BaseParameterTestCase
+from .util import BaseParameterTestCase
 
 
 class SelectToolParameterTestCase(BaseParameterTestCase):
